Add signTypedData methods
Per
[MetaMask/metamask-extension#7075](https://github.com/MetaMask/metamask-extension/issues/7075):

This adds all of the `signTypedData` methods currently supported by
MetaMask.

It has a breaking change:
- `signTypedData` previously was performing the `v3` behavior, but this
is very extra confusing (as we already have a version of this method
that defaults to `v1` in the main app), and we're not using this method in MetaMask, so I felt comfortable breaking it.

Does not really validate the signatures, merely enforces that the
correct methods from `eth-sig-util` are being used, which has the more
detailed behavioral tests in its own repository.

Composed in such a way that now `eth-hd-keyring` could inherit from this
and instantly gain these new methods (I believe, would want to add tests
to be sure).