Replace `withAppKeyOrigin` option with `getAppKeying`
The new `getAppKeyring` method will return a new SimpleKeyring created
with an app key derived from the provided origin. This serves as an
alternative to optionally augmenting each wallet method with a set of
options that may contain a `withAppKeyOrigin` value, which would be
used to generate a new wallet on-demand.

Moving app key generation to a single place makes it easier to limit
access to sensitive information, such as the primary private keys.
It also allows the consumer to use app keys more efficiently by using
the same wallet for successive operations, rather than generating a
new private key each time with `keccak`.