Standardize repo (#42)
- Add `.nvmrc` with `v10`
- Add `CODEOWNERS`
- `yarn import`
- Add `eslint` and `@metamask/eslint-config`; lint
- Update CircleCI config
  - Shadows recent library CircleCI configs
  - Removes `--ignore-scripts` flag from deps install script, as `scrypt` has required scripts
- Update dev dependencies to resolve security vulnerability
- Remove re-declaration of `decryptMessage` class method
  - The re-declaration was identical, so it didn't matter in practice

No functional changes were made, and no production dependencies were changed. It's possible that some of the syntactic changes per our lint rules are breaking for older versions of Node, so we may want to consider a major version bump to follow this. It should not affect us in any way.