Remove `ethereum-tx` package (#79)
This package was partially replaced by it's successor `@ethereumjs/tx`
in #61, but was left around for the unit tests to ensure the keyring
was compatible with both versions. Now that we have moved away from
using `ethereum-tx` in in MetaMask itself, we don't really need to
keep two versions of the same package around anymore.