Migrate tests from Mocha to Jest
The tests have been migrated from Mocha to Jest. Most of the changes
are the switch from `assert` to `expect`.

The "returns a signed tx object" test switched from checking the `raw`
property to using the `isSigned()` function because apparently `raw` is
always present and truthy, so the test didn't work in the first place.
The test description was updated to better reflect the transaction
object type being tested as well.

A redudant entry was removed from the `.gitignore` file as well, and
the headers re-ordered to be less misleading.