Add `@metamask/auto-changelog` (#75)
The `@metamask/auto-changelog` project has been added to help with
maintaining the changelog. An empty changelog has been created to
start with, and changelog validation has been added to CI.

The `repository` field in `package.json` had to be updated because the
URL was in a format that is not supported by
`@metamask/auto-changelog`.