Always return rejected Promise upon failure
All methods that return Promises now use the `async` keyword. This
simplifies the methods and ensures a consistent return value in the
case of an error.

For example, the method `getAppKeyAddress` would sometimes reject if it
encountered an error, and sometimes it would throw an error. Now it
will always reject, which is what you would typically expect of an
asynchronous function.