Remove unused `events` dependency (#86)
This dependency appears to be used, but in fact it is the Node.js
`events` API that gets used, because it always takes precedent over
package imports. This package was effectively ignored.

This package isn't required for browser compatibility anyway, as
bundlers like Webpack and browserify will substitute Node.js APIs for
browser-compatible implementations. That's how this has been working in
practice.