Make version-specific `signTypedData` methods private (#84)
* Make version-specific `signTypedData` methods private

The methods for a specific version of `signTypedData` are now private.
The `signTypedData` method accepts an `options` parameter that allows
using a specific version, and in practice this is the only method
called in MetaMask anyway.

* Remove unreachable default parameters

The `opts` parameter for the version-specific `signTypedData` functions
is now guaranteed to be truthy, because the functions are now private
and only called internally. The default value for `opts` has been
removed to improve our unit test branch coverage.