Remove `ethereumjs-wallet` (#88)
The `ethereumjs-wallet` package was replaced by a couple of simple
utility functions. The only thing we used the "Wallet" abstraction for
was generating a private key and getting the address that corresponds
to that private key.

This should behave identically to how it did before, except now we have
one less dependency. This is especially beneficial in this case because
`ethereumjs-wallet` had transitive dependencies that were inconvenient
for us (e.g. native dependencies and things that make LavaMoat
unimpressed).

Closes #21, #59