Update from `eth-sig-util@3` to `@metamask/eth-sig-util@4`
This should include no functional changes. The three internal functions
for calling each version of `signTypedData` have been removed, as
they're no longer needed. A condition was added to ensure invalid
versions are treated as "V1" so that the existing behaviour would be
preserved.

The test coverage for functions was lowered because there are now fewer
functions. Test coverage will be improved in a separate PR.